Update PHP requirements
As I want to guarantee that the code works - and keeps working - I want to run continuous integration tests on all supported versions. Since support for PHP 7.2 is about to end soon _and_ because there are no big changes or high priority bug fixes anticipated support for PHP 7.1 and PHP 7.2 can be dropped. For users that still run one of those versions the previous version will still work as expected. This also made upgrading PHPUnit easier as this allows for version 9.4 or higher instead of the need to juggle multiple versions.